refactor: Refatorando componente utilizado para proteger rotas privadas adicionando uma responsabilidade única para cada componente
