+++ src/layouts/ProtectedProae.tsx
@@ -1,0 +1,34 @@
+import SideBar from "@/components/SideBar/SideBar";
+import { Outlet, useLocation } from "react-router-dom";
+
+export default function ProtectedRouteProae() {
+  const location = useLocation();
+
+  const routesToNotRenderSideBar = ["/portal-proae/cadastro-edital"];
+  const shouldShowSideBar = !routesToNotRenderSideBar.includes(
+    location.pathname
+  );
+
+  return (
+    <div className="app-layout">
+      {shouldShowSideBar && (
+        <SideBar
+          homeIconRedirect={"/portal-proae/inscricoes"}
+          processIconRedirect={"/portal-proae/processos"}
+          docsIconRedirect={""}
+          configIconRedirect={"/portal-proae/configuracao"}
+          logoutIconRedirect={"/"}
+          logoutOnClick={() => {
+            console.log("logout logic for proae");
+          }}
+        />
+      )}
+
+      <main
+        className={`main-content ${shouldShowSideBar ? "with-sidebar" : "without-sidebar"}`}
+      >
+        <Outlet />
+      </main>
+    </div>
+  );
+}
